requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
cryptography==46.0.5
cymem==2.0.13
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
nltk==3.9.3
numpy==2.4.2
packaging==26.0
pdfminer.six==20260107
pillow==12.1.1
preshed==3.0.12
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMySQL==1.1.2
pytesseract==0.3.13
python-docx==1.2.0
python-multipart==0.0.22
regex==2026.2.28
requests==2.32.5
rich==14.3.3
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.0
shellingham==1.5.4
smart_open==7.5.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.48
srsly==2.5.2
starlette==0.52.1
thinc==8.3.10
threadpoolctl==3.6.0
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
wasabi==1.1.3
weasel==0.4.3
wrapt==2.1.1
